pyproject.toml
[project]
name = "temp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "fastapi>=0.116.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langgraph>=0.5.1",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "openai>=1.93.1",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "ruff>=0.12.2",
    "uvicorn>=0.35.0",
]
